Notification sediding using routes
