+++ Frontend/public/firebase-messaging-sw.js
@@ -1,5 +1,9 @@
+// import demo from "../src/assets/demo.js"
+
 importScripts("https://www.gstatic.com/firebasejs/8.10.0/firebase-app.js");
 importScripts("https://www.gstatic.com/firebasejs/8.10.0/firebase-messaging.js");
+
+// console.log(demo)
 
 const firebaseConfig = {
     apiKey: "AIzaSyBY-PXD09ZyOLdsjLVc3Aby6W43BlPdcEo",
@@ -19,6 +23,7 @@ messaging.onBackgroundMessage((payload) => {
         "[firebase-messaging-sw.js] Received background message ",
         payload
     );
+
     const notificationTitle = payload.notification.title;
     const notificationOptions = {
         body: payload.notification.body,
@@ -29,4 +34,5 @@ messaging.onBackgroundMessage((payload) => {
     };
 
     self.registration.showNotification(notificationTitle, notificationOptions);
+
 });
