+++ Frontend/src/assets/demo.js
@@ -1,0 +1,3 @@
+const value = 10;
+
+export default value;
